add transition gradient for token dropdown search and refactor code
