+++ src/pages/address/components/DropDownTokenRow.tsx
@@ -1,40 +1,6 @@
 import React from "react";
 import NumericFormat from "@components/commons/NumericFormat";
-import BTC from "@components/icons/BTC";
-import dBCH from "@components/icons/dBCH";
-import DFI from "@components/icons/DFI";
-import dBTC from "@components/icons/dBTC";
-import dDFI from "@components/icons/dDFI";
-import dETH from "@components/icons/dETH";
-import dLTC from "@components/icons/dLTC";
-import dUSDT from "@components/icons/dUSDT";
-import dUSDC from "@components/icons/dUSDC";
-import dUSD from "@components/icons/dUSD";
-import DefaultToken from "@components/icons/DefaultToken";
-
-interface MappingProps {
-  size?: number;
-  symbol?: string;
-}
-
-const mapping: Record<string, ({ size, symbol }: MappingProps) => JSX.Element> =
-  {
-    _UTXO: DFI,
-    DFI,
-    "DFI (UTXO)": DFI,
-    "DFI (Token)": DFI,
-    BTC,
-    dBCH,
-    dBTC,
-    dDFI,
-    dETH,
-    dLTC,
-    dUSDT,
-    dUSDC,
-    dUSD,
-    DUSD: dUSD,
-    DefaultToken,
-  };
+import { getTokenIcon } from "../../../shared/getTokenIcon";
 
 interface DropDownTokenRowProps {
   value: number;
@@ -45,7 +11,7 @@ export default function DropDownTokenRow({
   value,
   symbol,
 }: DropDownTokenRowProps) {
-  const Icon = getIcon(symbol);
+  const Icon = getTokenIcon(symbol);
   return (
     <div className="pb-6 flex flex-row items-center">
       <div className="flex flex-row gap-x-2 grow items-center">
@@ -63,12 +29,4 @@ export default function DropDownTokenRow({
       </div>
     </div>
   );
-}
-
-function getIcon(symbol: string): (size: MappingProps) => JSX.Element {
-  let Icon = mapping[symbol];
-  if (Icon === undefined) {
-    Icon = mapping.DefaultToken;
-  }
-  return Icon;
 }
